move method in serviceProvider to sessionManager
