+++ server/src/main/java/org/apache/iotdb/db/protocol/mqtt/PublishHandler.java
@@ -20,9 +20,9 @@ package org.apache.iotdb.db.protocol.mqtt;
 import org.apache.iotdb.db.conf.IoTDBConfig;
 import org.apache.iotdb.db.metadata.path.PartialPath;
 import org.apache.iotdb.db.qp.physical.crud.InsertRowPlan;
+import org.apache.iotdb.db.query.control.SessionManager;
 import org.apache.iotdb.db.service.IoTDB;
 import org.apache.iotdb.db.service.basic.BasicOpenSessionResp;
-import org.apache.iotdb.db.service.basic.ServiceProvider;
 import org.apache.iotdb.service.rpc.thrift.TSProtocolVersion;
 import org.apache.iotdb.service.rpc.thrift.TSStatus;
 
@@ -42,7 +42,7 @@ import java.util.List;
 /** PublishHandler handle the messages from MQTT clients. */
 public class PublishHandler extends AbstractInterceptHandler {
 
-  private final ServiceProvider serviceProvider = IoTDB.serviceProvider;
+  private final SessionManager SESSION_MANAGER = SessionManager.getInstance();
   private long sessionId;
 
   private static final Logger LOG = LoggerFactory.getLogger(PublishHandler.class);
@@ -66,11 +66,12 @@ public class PublishHandler extends AbstractInterceptHandler {
   public void onConnect(InterceptConnectMessage msg) {
     try {
       BasicOpenSessionResp basicOpenSessionResp =
-          serviceProvider.openSession(
-              msg.getUsername(),
-              new String(msg.getPassword()),
-              ZoneId.systemDefault().toString(),
-              TSProtocolVersion.IOTDB_SERVICE_PROTOCOL_V3);
+          SessionManager.getInstance()
+              .openSession(
+                  msg.getUsername(),
+                  new String(msg.getPassword()),
+                  ZoneId.systemDefault().toString(),
+                  TSProtocolVersion.IOTDB_SERVICE_PROTOCOL_V3);
       sessionId = basicOpenSessionResp.getSessionId();
     } catch (TException e) {
       throw new RuntimeException(e);
@@ -79,7 +80,7 @@ public class PublishHandler extends AbstractInterceptHandler {
 
   @Override
   public void onDisconnect(InterceptDisconnectMessage msg) {
-    serviceProvider.closeSession(sessionId);
+    SESSION_MANAGER.closeSession(sessionId);
   }
 
   @Override
@@ -119,11 +120,11 @@ public class PublishHandler extends AbstractInterceptHandler {
                 event.getTimestamp(),
                 event.getMeasurements().toArray(new String[0]),
                 event.getValues().toArray(new String[0]));
-        TSStatus tsStatus = serviceProvider.checkAuthority(plan, sessionId);
+        TSStatus tsStatus = SESSION_MANAGER.checkAuthority(plan, sessionId);
         if (tsStatus != null) {
           LOG.warn(tsStatus.message);
         } else {
-          status = serviceProvider.executeNonQuery(plan);
+          status = IoTDB.serviceProvider.executeNonQuery(plan);
         }
       } catch (Exception e) {
         LOG.warn(
