+++ server/src/main/java/org/apache/iotdb/db/query/control/SessionManager.java
@@ -20,14 +20,20 @@ package org.apache.iotdb.db.query.control;
 
 import org.apache.iotdb.commons.conf.IoTDBConstant;
 import org.apache.iotdb.db.auth.AuthException;
+import org.apache.iotdb.db.auth.AuthorityChecker;
 import org.apache.iotdb.db.auth.authorizer.BasicAuthorizer;
 import org.apache.iotdb.db.auth.authorizer.IAuthorizer;
+import org.apache.iotdb.db.conf.OperationType;
 import org.apache.iotdb.db.exception.StorageEngineException;
 import org.apache.iotdb.db.mpp.common.SessionInfo;
+import org.apache.iotdb.db.qp.physical.PhysicalPlan;
+import org.apache.iotdb.db.qp.physical.sys.AuthorPlan;
 import org.apache.iotdb.db.query.dataset.UDTFDataSet;
 import org.apache.iotdb.db.service.basic.BasicOpenSessionResp;
+import org.apache.iotdb.rpc.RpcUtils;
 import org.apache.iotdb.rpc.TSStatusCode;
 import org.apache.iotdb.service.rpc.thrift.TSProtocolVersion;
+import org.apache.iotdb.service.rpc.thrift.TSStatus;
 import org.apache.iotdb.tsfile.read.query.dataset.QueryDataSet;
 
 import org.apache.thrift.TException;
@@ -41,6 +47,8 @@ import java.util.TimeZone;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CopyOnWriteArraySet;
 import java.util.concurrent.atomic.AtomicLong;
+
+import static org.apache.iotdb.db.utils.ErrorHandlingUtils.onNPEOrUnexpectedException;
 
 public class SessionManager {
   private static final Logger LOGGER = LoggerFactory.getLogger(SessionManager.class);
@@ -76,23 +84,6 @@ public class SessionManager {
 
   protected SessionManager() {
     // singleton
-  }
-
-  public Long getCurrSessionId() {
-    return currSessionId.get();
-  }
-
-  public void removeCurrSessionId() {
-    currSessionId.remove();
-  }
-
-  public TimeZone getCurrSessionTimeZone() {
-    if (getCurrSessionId() != null) {
-      return TimeZone.getTimeZone(SessionManager.getInstance().getZoneId(getCurrSessionId()));
-    } else {
-      // only used for test
-      return TimeZone.getTimeZone("+08:00");
-    }
   }
 
   public BasicOpenSessionResp openSession(
@@ -154,12 +145,54 @@ public class SessionManager {
     return openSessionResp.sessionId(sessionId);
   }
 
+  public BasicOpenSessionResp openSession(
+      String username, String password, String zoneId, TSProtocolVersion tsProtocolVersion)
+      throws TException {
+    return openSession(
+        username, password, zoneId, tsProtocolVersion, IoTDBConstant.ClientVersion.V_0_12);
+  }
+
   public boolean closeSession(long sessionId) {
     AUDIT_LOGGER.info("Session-{} is closing", sessionId);
-
-    removeCurrSessionId();
-
+    currSessionId.remove();
     return SessionTimeoutManager.getInstance().unregister(sessionId);
+  }
+
+  public TSStatus closeOperation(
+      long sessionId,
+      long queryId,
+      long statementId,
+      boolean haveStatementId,
+      boolean haveSetQueryId) {
+    if (!checkLogin(sessionId)) {
+      return RpcUtils.getStatus(
+          TSStatusCode.NOT_LOGIN_ERROR,
+          "Log in failed. Either you are not authorized or the session has timed out.");
+    }
+
+    if (AUDIT_LOGGER.isDebugEnabled()) {
+      AUDIT_LOGGER.debug(
+          "{}: receive close operation from Session {}",
+          IoTDBConstant.GLOBAL_DB_NAME,
+          currSessionId);
+    }
+
+    try {
+      if (haveStatementId) {
+        if (haveSetQueryId) {
+          this.closeDataset(statementId, queryId);
+        } else {
+          this.closeStatement(sessionId, statementId);
+        }
+        return RpcUtils.getStatus(TSStatusCode.SUCCESS_STATUS);
+      } else {
+        return RpcUtils.getStatus(
+            TSStatusCode.CLOSE_OPERATION_ERROR, "statement id not set by client.");
+      }
+    } catch (Exception e) {
+      return onNPEOrUnexpectedException(
+          e, OperationType.CLOSE_OPERATION, TSStatusCode.CLOSE_OPERATION_ERROR);
+    }
   }
 
   /**
@@ -168,7 +201,7 @@ public class SessionManager {
    * @return true: If logged in; false: If not logged in
    */
   public boolean checkLogin(long sessionId) {
-    boolean isLoggedIn = getUsername(sessionId) != null;
+    boolean isLoggedIn = sessionIdToUsername.get(sessionId) != null;
     if (!isLoggedIn) {
       LOGGER.info("{}: Not login. ", IoTDBConstant.GLOBAL_DB_NAME);
     } else {
@@ -270,6 +303,49 @@ public class SessionManager {
       } catch (Exception e) {
         LOGGER.warn("Error occurred while releasing query resource: ", e);
       }
+    }
+  }
+
+  public boolean checkAuthorization(PhysicalPlan plan, String username) throws AuthException {
+    if (!plan.isAuthenticationRequired()) {
+      return true;
+    }
+
+    String targetUser = null;
+    if (plan instanceof AuthorPlan) {
+      targetUser = ((AuthorPlan) plan).getUserName();
+    }
+    return AuthorityChecker.check(
+        username, plan.getAuthPaths(), plan.getOperatorType(), targetUser);
+  }
+
+  public TSStatus checkAuthority(PhysicalPlan plan, long sessionId) {
+    try {
+      if (!checkAuthorization(plan, sessionIdToUsername.get(sessionId))) {
+        return RpcUtils.getStatus(
+            TSStatusCode.NO_PERMISSION_ERROR,
+            "No permissions for this operation " + plan.getOperatorType());
+      }
+    } catch (AuthException e) {
+      LOGGER.warn("meet error while checking authorization.", e);
+      return RpcUtils.getStatus(TSStatusCode.UNINITIALIZED_AUTH_ERROR, e.getMessage());
+    } catch (Exception e) {
+      return onNPEOrUnexpectedException(
+          e, OperationType.CHECK_AUTHORITY, TSStatusCode.EXECUTE_STATEMENT_ERROR);
+    }
+    return null;
+  }
+
+  public Long getCurrSessionId() {
+    return currSessionId.get();
+  }
+
+  public TimeZone getCurrSessionTimeZone() {
+    if (getCurrSessionId() != null) {
+      return TimeZone.getTimeZone(SessionManager.getInstance().getZoneId(getCurrSessionId()));
+    } else {
+      // only used for test
+      return TimeZone.getTimeZone("+08:00");
     }
   }
